Make BackPressedCallback public
